+++ Assets/Script/util.cs
@@ -12,39 +12,39 @@ public class Util
     {
         return x + y * width;
     }
-    private static long IntToNat(int z)
+    private static uint IntToNat(int z)
     {
-        return z >= 0 ? (long)z * 2 : (long)(-z) * 2 - 1;
+        return z >= 0 ? (uint)z * 2 : (uint)(-z) * 2 - 1;
     }
 
-    private static int NatToInt(long n)
+    private static int NatToInt(uint n)
     {
         if (n % 2 == 0) return (int)(n / 2);
         else return -(int)((n + 1) / 2);
     }
 
-    private static long CantorPair(long k, long l)
+    private static uint CantorPair(uint k, uint l)
     {
         return ((k + l) * (k + l + 1)) / 2 + l;
     }
 
-    private static (long, long) CantorUnpair(long z)
+    private static (uint, uint) CantorUnpair(uint z)
     {
-        long w = (long)math.floor((math.sqrt(8 * (double)z + 1) - 1) / 2);
-        long t = (w * (w + 1)) / 2;
-        long l = z - t;
-        long k = w - l;
+        uint w = (uint)math.floor((math.sqrt(8 * (double)z + 1) - 1) / 2);
+        uint t = (w * (w + 1)) / 2;
+        uint l = z - t;
+        uint k = w - l;
         return (k, l);
     }
 
-    public static long Encode(int n1, int n2)
+    public static uint Encode(int n1, int n2)
     {
-        long g1 = IntToNat(n1);
-        long g2 = IntToNat(n2);
+        uint g1 = IntToNat(n1);
+        uint g2 = IntToNat(n2);
         return CantorPair(g1, g2);
     }
 
-    public static (int, int) Decode(long m)
+    public static (int, int) Decode(uint m)
     {
         var (g1, g2) = CantorUnpair(m);
         int n1 = NatToInt(g1);
